requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
python-docx==0.8.11
pandas==2.1.3
openpyxl==3.1.2
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
sendgrid==6.10.0
python-dotenv==1.0.0
docx2pdf==0.1.8
flower==0.9.4
python-celery==5.3.4
kombu==5.3.4
billiard==4.1.1
